requirements.txt
slack_bolt
openai
python-dotenv